backend/requirements.txt
# Backend Dependencies for Upload Pipeline
# Add these to your existing requirements.txt

minio>=7.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
